% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aritmetica}
\alias{aritmetica}
\title{Titulo da minha funcao}
\usage{
aritmetica(a, b, op = "soma")
}
\arguments{
\item{a}{descricao do parametro a}

\item{b}{descricao do parametro b}

\item{op}{funcao de calculo}
}
\value{
diga o que a funcao vai retornar, um numero / uma tibble / um vetor
}
\description{
Pula uma linha após o título e acrescenta funcoes adicionais
}
