% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_var.R
\name{coef_var}
\alias{coef_var}
\title{Função Coeficiente de Variação}
\usage{
coef_var(x)
}
\arguments{
\item{x}{é um vetor numérico}
}
\description{
Função criada por Marcos V C Vital, da Academia do R
Criada em R 4.0.3
}
\details{
Calcula o coeficiente de variação, expresso em porcentagem
}
\examples{

library (dplyr)
library (readxl)
dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosPop.xlsx") \%>\%
select (altura)

coef_var (dados$altura)

}
