% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefVar.R
\name{coefVar}
\alias{coefVar}
\title{Função Coeficiente de Variação}
\usage{
coefVar(x)
}
\arguments{
\item{x}{é um vetor numérico}
}
\description{
Calcula o coeficiente de variação, expresso em porcentagem
}
\examples{

library (readxl)
dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosPop.xlsx")

coefVar (dados$altura)

}
\author{
Petronio Fagundes de Oliveira Filho
}
