% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercicio.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontrar pior ano do Brasileirao para um time}
\usage{
encontrar_pior_ano_time(time = "Cruzeiro")
}
\arguments{
\item{time}{Nome de um time como na coluna 'quem_ganhou' de
\link{partidas_brasileirao} ("Cruzeiro por padrao")}
}
\value{
Uma tibble com tres colunas
}
\description{
Usando a base Brasileirao, encontrar pior ano
(menor numero de vitorias) de um time na serie historica
}
\examples{
encontrar_pior_ano_time("Corinthians")

}
