% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erroLinha.R
\name{erroLinha}
\alias{erroLinha}
\title{Função gráfico de linha com erro}
\usage{
erroLinha(d = NULL, x = NULL, y = NULL, erro = NULL)
}
\arguments{
\item{d}{é um dataframe com os dados}

\item{x}{é uma variável categórica (fator)}

\item{y}{é uma variável númerica}

\item{erro}{é a margem de erro usada, ver exemplo}
}
\description{
Esta função desenha um gráfico de linha com erro
}
\examples{
library (readxl)
library (dplyr)
library (ggplot2)

smoke <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosFumo.xlsx")

smoke$fumo <- factor(smoke$fumo,
                    levels = c(1, 2, 3, 4),
                    labels = c("não", "leve", "moderado", "pesado"))

smoke1 <- smoke \%>\%
dplyr::group_by(fumo) \%>\%
 summarise(n = n(),
           media = mean(pesoRN),
           dp = sd (pesoRN),
           me = dp/sqrt(n) * qt(1 - (0.05/2), n - 1))

erroLinha (d = smoke1, y = "media", x = "fumo", erro = "me") +
          labs(x="Tabagismo", y="Peso ao nascer")

}
\author{
Petronio Fagundes de Oliveira Filho
}
