% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erroPad.R
\name{erroPad}
\alias{erroPad}
\title{Erro Padrão da Média}
\usage{
erroPad(x)
}
\arguments{
\item{x}{Vetor numérico.}
}
\value{
Valor numérico: \eqn{s / \sqrt{n}}.
}
\description{
Calcula o erro padrão da média amostral. Mede a variação esperada da
média amostral em torno da média populacional, sendo a base para a
construção de intervalos de confiança.
}
\examples{
pesoRN <- c(3786, 3891, 2761, 2891, 3880, 2780, 3188, 3387,
            3394, 3458, 3992, 2897, 3284, 3412, 3527, 2732,
            3373)
erroPad(pesoRN)

}
\author{
Petronio Fagundes de Oliveira Filho
}
