% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-filtrar-videos-por-rel.R
\name{filtrar_rel}
\alias{filtrar_rel}
\title{Filtrar vídeos pelas métricas de relevância}
\usage{
filtrar_rel(filtro_likes, filtro_comen)
}
\arguments{
\item{filtro_likes}{quantidade de likes}

\item{filtro_comen}{quantidade de comentários}
}
\value{
um data frame
}
\description{
Essa função tem como objetivo simplificar a triagem dos vídeos da base
do canal Brasil Paralelo, com base em dois critérios que enumerei
para representar o grau de relevância dos vídeos, o número de likes
e o número de comentários dos vídeos. Os critérios devem ser designados
conforme a numeração escolhida, e colocados nos argumentos filtro_likes = x
para o número de likes desejados e filtro_comen = x para o número de
comentários estabelecido.
}
\examples{
filtrar_rel(filtro_likes = 1000, filtro_comen = 200)
}
