% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_buscar_videos_por_categorias.R
\name{filtrar_videos_por_categorias}
\alias{filtrar_videos_por_categorias}
\title{Filtrar vídeos pelas categorias}
\usage{
filtrar_videos_por_categorias(filtro_categorias, filtro_mes, filtro_ano)
}
\arguments{
\item{filtro_categorias}{filtra as categorias, categorias disponíveis "Educação", "Política", "Entretenimento" e "Diversos".}

\item{filtro_mes}{filtra os vídeos com base no mês.}

\item{filtro_ano}{filtra o vídeo com base no ano.}
}
\value{
um data frame
}
\description{
Essa função tem como objetivo filtrar a base de vídeos extraída do canal
Brasil Paralelo a partir do critério de categoria, estabelecida conforme
a classificação do YouTube, junto aos critérios de mês e ano do vídeo
em questão.
}
\examples{
filtrar_videos_por_categorias(filtro_categorias = "Educação", filtro_mes = 02, filtro_ano = 2022)
}
