% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{fun}
\alias{fun}
\title{funcao para somar dois numeros}
\usage{
fun(a, b)
}
\arguments{
\item{a}{um numero para somar}

\item{b}{um numero para somar}
}
\value{
numero
}
\description{
funcao para somar dois numeros
}
\examples{
fun(2, 2)
}
