% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{futebol}
\alias{futebol}
\title{Resultados do Brasileirão}
\format{
7226 linhas e 8 colunas
\describe{
\item{col1}{Descrição da coluna 1}
\item{col2}{Descrição da coluna 2}
...
}
}
\source{
Pacote Brasileirao por William Amorim
}
\usage{
futebol
}
\description{
Dataset de resultados dos jogos do Brasileirão
}
\keyword{datasets}
