% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histNormal.R
\name{histNormal}
\alias{histNormal}
\title{Histograma com curva normal sobreposta}
\usage{
histNormal(dados = NULL, x = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{dados}{Dataframe com os dados.}

\item{x}{Vetor numérico base para o histograma.}

\item{xlab}{Rótulo do eixo x.}

\item{ylab}{Rótulo do eixo y.}
}
\value{
Um objeto ggplot.
}
\description{
Constrói um histograma e sobrepõe a curva da distribuição normal
parametrizada pela média e desvio padrão dos dados.
}
\examples{
pesoRN <- c(3340, 3345, 3750, 3650, 3220, 4070, 3380, 3970,
            3060, 3180, 2865, 2815, 3245, 2051, 2630)
sexo <- c(2, 2, 2, 1, 1, 1, 2, 1, 1, 1, 2, 2, 1, 1, 2)
neonato <- data.frame(pesoRN, sexo)
histNormal(dados = neonato,
           x = neonato$pesoRN,
           ylab = "Densidade de Probabilidade",
           xlab = "Peso ao nascer (g)")

}
\author{
Petronio Fagundes de Oliveira Filho
}
