% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ler_varios_csv}
\alias{ler_varios_csv}
\title{Esta funcao importa varios arquivos csv de
uma pasta}
\usage{
ler_varios_csv(pasta)
}
\arguments{
\item{pasta}{uma string referente a pasta onde os csvs serao procurados}
}
\value{
Uma tibble
}
\description{
Esta funcao importa varios arquivos csv de
uma pasta
}
\examples{
 \dontrun{
 ler_varios_csv("caminho_ate_pasta")
 }
}
