% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meuxy.R
\name{meuxy}
\alias{meuxy}
\title{Faz um Grafico Personalizado}
\usage{
meuxy(x, ...)
}
\arguments{
\item{x}{Uma formula do lattice}

\item{...}{Outros argumentos passados para a funcao
    \code{\link[lattice]{xyplot}}}
}
\value{
Um grafico xyplot do lattice com \code{pch = 4} e \code{col =
    1}.
}
\description{
Uma (incrivel) funcao para fazer um xyplot
    personalizado. Usando o pacote lattice.
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
    sua funcao (se necessario), ou para detalhar melhor como
    utilizar determinados argumentos.
}
\examples{
meuxy(1:10 ~ 1:10)
}
\author{
Fernando Mayer
}
\seealso{
\code{\link[lattice]{xyplot}}
}

