% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mtcars_limpar}
\alias{mtcars_limpar}
\title{mtcars_limpar}
\format{
Uma lista que descreve as colunas:
\describe{
\item{mpg}{Descrição da coluna 1}
\item{cyl}{Descrição da coluna 2}
\item{disp}{Descrição da coluna 2}
\item{hp}{Descrição da coluna 2}
\item{drat}{Descrição da coluna 2}
\item{wt}{Descrição da coluna 2}
\item{qsec}{Descrição da coluna 2}
\item{vs}{Descrição da coluna 2}
\item{am}{Descrição da coluna 2}
\item{gear}{Descrição da coluna 2}
\item{carb}{Descrição da coluna 2}

}
}
\usage{
mtcars_limpar
}
\description{
Minha base mtcars
}
\keyword{datasets}
