% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Partidas do Brasileirao}
\format{
A base tem 7266 linhas e 8 colunas:
\describe{
\item{temporada}{Ano do Brasileirao}
\item{data}{Data da partida}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
partidas_brasileirao
}
\description{
Serie historica de todas as partidas do Brasileirao.
}
\keyword{datasets}
