% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bp.R
\name{plot_bp}
\alias{plot_bp}
\title{Função desenha boxplots com teste t de Student}
\usage{
plot_bp(df, var.x, var.y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var.x}{é uma variável categórica dicotômica (fator)}

\item{var.y}{é uma variável númerica}
}
\description{
Esta função desenha os boxplots comparando dois grupos independentes e mostra
o resultado do teste t de Student
}
\examples{

dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosPop.xlsx")
str(dados)
dados$pop <- as.factor(dados$pop)

by (data = dados$altura, INDICES = dados$pop, FUN = shapiro.test)

leveneTest(altura ~ pop, center = mean, data = dados)

plot_bp (df = dados, var.x = pop, var.y = altura)

}
\author{
Petronio Fagundes de Oliveira Filho
}
