% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bpW.R
\name{plot_bpW}
\alias{plot_bpW}
\title{Função que desenha boxplots comparando duas médias com teste de Wilcox}
\usage{
plot_bpW(df, var.x, var.y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var.x}{é uma variável categórica dicotômica}

\item{var.y}{é uma variável númerica}
}
\description{
Esta calcula um teste de Wilcox para duas médias independentes e desenha os
boxplots com o resultado do teste
}
\examples{

cirurgia <- read_excel ("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosCirurgia.xlsx")

cirurgia$infec <- as.factor(cirurgia$infec)

plot_bpW(df = cirurgia, var.x = infec, var.y = tempohosp)

}
