% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBpW.R
\name{plotBpW}
\alias{plotBpW}
\title{Função desenha boxplots com teste de Wilcoxon (Mann-Whitney)}
\usage{
plotBpW(df, var.x, var.y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var.x}{é uma variável categórica dicotômica}

\item{var.y}{é uma variável númerica}
}
\description{
Esta função desenha os boxplots comparando dois grupos independentes e mostra
o resultado do teste de Wilcoxon
}
\examples{

cirurgia <- read_excel ("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosCirurgia.xlsx")

cirurgia$infec <- as.factor(cirurgia$infec)

plotBpW(df = cirurgia, var.x = infec, var.y = tempohosp)

}
\author{
Petronio Fagundes de Oliveira Filho
}
