% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReg.R
\name{plotReg}
\alias{plotReg}
\title{Função desenha gráfico de dispersão}
\usage{
plotReg(df, var_x, var_y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var_x}{é a variável preditora (independente) numérica}

\item{var_y}{é uma variável desfecho (dependente) númerica}
}
\description{
Esta função desenha um gráfico de dispersão com reta de ajuste com IC95% e
equação da reta
}
\examples{

library (readxl)
library (ggplot2)
library (ggpubr)

# Dados de crianças até 36 meses idade
dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosReg.xlsx")

shapiro.test(dados$idade)
shapiro.test(dados$comp)

plotReg (df = dados, var_x = idade, var_y = comp) +
        labs(x="Idade (meses)", y="Comprimento (cm)")

}
\author{
Petronio Fagundes de Oliveira Filho
}
