% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riscoRelativo.R
\name{riscoRelativo}
\alias{riscoRelativo}
\title{Função Risco Relativo}
\usage{
riscoRelativo(tabela, alpha = 0.05, referencerow = 2)
}
\arguments{
\item{tabela}{Tabela de tabulação cruzada 2 x 2.}

\item{alpha}{Nível de significância (padrão: 0.05).}

\item{referencerow}{Linha de referência da tabela (padrão: 2).}
}
\description{
Calcula o Risco Relativo a partir de uma tabela 2 x 2.
}
\details{
Crie a tabela 2 x 2 antes de usar a função:

tabela <- matrix(c(valor1, valor2, valor3, valor4), nrow = 2, byrow = TRUE)

colnames(tabela) <- c("Doenca", "Sem doenca")

rownames(tabela) <- c("Exposto", "Nao exposto")
}
\author{
Petronio Fagundes de Oliveira Filho
}
