% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcaosomar.R
\name{somar}
\alias{somar}
\title{Soma dois elementos}
\usage{
somar(x, y)
}
\arguments{
\item{x}{primeiro elemento}

\item{y}{segundo elemento}
}
\description{
soma dois elementos
}
\examples{
x <-1
y <-2
somar(x,y)

}
