% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefVar.R
\name{coefVar}
\alias{coefVar}
\title{Função Coeficiente de Variação}
\usage{
coefVar(x)
}
\arguments{
\item{x}{é um vetor numérico}
}
\description{
Calcula o coeficiente de variação, expresso em porcentagem
}
\examples{
pesoRN <- c(3340, 3345, 3750, 3650, 3220, 4070, 3380, 3970,
           3060, 3180, 2865, 2815, 3245, 2051, 2630)

sexo <- c(2, 2, 2, 1, 1, 1, 2, 1, 1, 1, 2, 2, 1, 1, 2)

neonato <- data.frame(pesoRN, sexo)

neonato$sexo <- as.factor(neonato$sexo)

coefVar (neonato$pesoRN)

}
\author{
Petronio Fagundes de Oliveira Filho
}
