% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCohen.R
\name{dCohen}
\alias{dCohen}
\title{Função d de Cohen}
\usage{
dCohen(n1, n2, media1, media2, s1, s2)
}
\arguments{
\item{n1}{é tamanho do grupo 1}

\item{n2}{é tamanho do grupo 2}

\item{media1}{é a média do grupo 1}

\item{media2}{é a média do grupo 2}

\item{s1}{é desvio padrão do grupo 1}

\item{s2}{é desvio padrão do grupo 2}
}
\value{
O valor do d de Cohen e uma interpretação da sua magnitude de acordo
com Cohen (1988)
}
\description{
Calcula o d de Cohen, medida usada para ponderar a magnitude do efeito
na comparação entre duas médias
}
\examples{

library (dplyr)
library (readxl)
dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosPop.xlsx")
dados$pop <- factor(dados$pop)

media <- aggregate(dados$altura, list(dados$pop), FUN=mean)
s <- aggregate(dados$altura, list(dados$pop), FUN=sd)
n <- dados \%>\% count(pop, sort = TRUE)

s1 <- s$x [1]
s2 <- s$x [2]
n1 <- n$n [1]
n2 <- n$n [2]

media1 <- media$x [1]
media2 <- media$x [2]

dCohen(n1, n2, media1, media2, s1, s2)

}
\author{
Petronio Fagundes de Oliveira Filho
}
