% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erroPad.R
\name{erroPad}
\alias{erroPad}
\title{Função Erro Padrão}
\usage{
erroPad(x)
}
\arguments{
\item{x}{é um vetor numérico}
}
\description{
Calcula o erro padrão, medida de variação de uma média amostral em relação à
média da população. Ajuda a verificar a confiabilidade da média amostral, pois
atrávés dele podemos estimar o intervalo de confiança para a média.
}
\examples{

library (readxl)
dados <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosPop.xlsx")

erroPad (dados$altura)

}
\author{
Petronio Fagundes de Oliveira Filho
}
