% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_normal.R
\name{hist_normal}
\alias{hist_normal}
\title{Função de curva normal sobreposta ao histograma}
\usage{
hist_normal(dados = NULL, x = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{dados}{é um dataframe com os dados}

\item{x}{é uma variável numérica base para o histograma}

\item{xlab}{é o rotulo do eixo x}

\item{ylab}{é o rótulo do eixo y}
}
\description{
Esta função constroi um histograma e coloca uma curva normal sobreposta
}
