% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histNormal.R
\name{histNormal}
\alias{histNormal}
\title{Função curva normal sobreposta ao histograma}
\usage{
histNormal(dados = NULL, x = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{dados}{é um dataframe com os dados}

\item{x}{é uma variável numérica base para o histograma}

\item{xlab}{é o rotulo do eixo x}

\item{ylab}{é o rótulo do eixo y}
}
\description{
Esta função constroi um histograma e coloca uma curva normal sobreposta
}
\examples{
library (readxl)
library (dplyr)

mater15 <- read_excel("C:/Users/petro/Dropbox/Estatística/Bioestatística usando o R/dadosMater15.xlsx") \%>\%
         select(pesoRN)

histNormal(dados = mater15, x = mater15$pesoRN, ylab = "Peso ao nascer (g)", xlab = "Recém-nascidos")

}
\author{
Petronio Fagundes de Oliveira Filho
}
