% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsRatio.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Função Odds Ratio}
\usage{
oddsRatio(tabela, alpha = 0.05, referencerow = 2, quiet = FALSE)
}
\arguments{
\item{tabela}{é uma tabela de tabulação cruzada. A tabela 2 x 2 deve ser criada
antes e nomeada como tabela. Ver Details}

\item{alpha}{é o nível de significância, padrão é 0.05}

\item{referencerow}{é a linha da tabela que será a referência, o padrão é 2}

\item{quiet}{argumento lógico, indicando se a saída detalhada deve ser
suprimida. O padrão é FALSE}
}
\description{
Esta função calcula a Odds Ratio a partir de uma tabela 2 x 2
}
\details{
<<<<<<< HEAD
<<<<<<< HEAD
Basta criar uma tabela 2 x 2 usar no modelo. Nomeia-a como quiser.
=======
=======
>>>>>>> dba9824e23d8686f7459f68f074895a0e40f2325
Basta criar uma tabela 2 x 2 para calcular RR, usando o formato:

tabela <- matrix(c(valor1, valor2, valor3, valor4), nrow=2, byrow=TRUE)

colnames(tabela) <- c("Doença","Sem doença")

rownames(tabela) <- c("Exposto","Não exposto")

print(tabela)

Usar a tabela no modelo
<<<<<<< HEAD
>>>>>>> dba9824e23d8686f7459f68f074895a0e40f2325
=======
>>>>>>> dba9824e23d8686f7459f68f074895a0e40f2325
}
\author{
Petronio Fagundes de Oliveira Filho
}
