% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bp.R
\name{plot_bp}
\alias{plot_bp}
\title{Função de teste t com boxplots}
\usage{
plot_bp(df, x_col, y_col)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{x_col}{é uma variável categórica dicotômica}

\item{y_col}{é uma variável númerica}
}
\description{
Esta calcula um teste t e desenha os boxplots com o resultado do teste
}
