% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBpW.R
\name{plotBpW}
\alias{plotBpW}
\title{Função desenha boxplots com teste de Wilcoxon (Mann-Whitney)}
\usage{
plotBpW(df, var.x, var.y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var.x}{é uma variável categórica dicotômica}

\item{var.y}{é uma variável númerica}
}
\description{
Esta função desenha os boxplots comparando dois grupos independentes e mostra
o resultado do teste de Wilcoxon
}
\examples{

library (car)
tempohosp <- c(58, 11, 153,  30,  19,  19,  23,  37,  44, 117,
              109,  15,  30,  14,  50,  17,  22,38,  13,  27)
infec <- c('não', 'não', 'sim', 'sim', 'sim', 'não', 'não', 'sim', 'não', 'não',
          'não', 'não', 'não', 'não', 'sim', 'não', 'não', 'sim', 'não', 'não')

cirurgia <- data.frame(tempohosp, infec)

cirurgia$infec <- as.factor(cirurgia$infec)

by (data = cirurgia$tempohosp, INDICES = cirurgia$infec, FUN = shapiro.test)

leveneTest(tempohosp ~ infec, center = median, data = cirurgia)

plotBpW(df = cirurgia, var.x = infec, var.y = tempohosp)

}
\author{
Petronio Fagundes de Oliveira Filho
}
