% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReg.R
\name{plotReg}
\alias{plotReg}
\title{Função desenha gráfico de dispersão}
\usage{
plotReg(df, var_x, var_y)
}
\arguments{
\item{df}{é um dataframe com os dados}

\item{var_x}{é a variável preditora (independente) numérica}

\item{var_y}{é uma variável desfecho (dependente) númerica}
}
\description{
Esta função desenha um gráfico de dispersão com reta de ajuste com IC95% e
equação da reta
}
\examples{
library (ggplot2)
library (ggpubr)

idade <- c(33, 24, 34, 35, 33, 25, 31, 28, 30, 34,
          24, 30, 35, 27, 32, 19, 26, 26, 26, 36)

comp <-  c(98.0, 90.5, 97.5, 97.0, 96.0, 89.5, 97.0, 91.5, 92.0, 99.0,
          90.0, 96.0, 100.0, 87.0, 93.5, 82.0, 92.0, 89.0, 87.5, 100.0)

dados <-  data.frame(idade, comp)

shapiro.test(dados$idade)
shapiro.test(dados$comp)

plotReg (df = dados, var_x = idade, var_y = comp) +
 labs(x="Idade (meses)", y="Comprimento (cm)")

}
\author{
Petronio Fagundes de Oliveira Filho
}
