% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{resumo}
\alias{resumo}
\title{Função para calcular medidas resumidoras com gráficos}
\usage{
resumo(x, lab)
}
\arguments{
\item{x}{é um vetor de uma variável numérica}

\item{lab}{é o rótulo da variável númerica, escrita entre aspas}
}
\description{
Esta função calcula medias resumidoras (média, desvio padrão, mediana, P25 e
P75), entregando junto um histograma com curva normal sobreposta e boxplot de
uma variável numérica
}
\author{
Petronio Fagundes de Oliveira Filho
}
