% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riscoRelativo.R
\name{RiscoRelativo}
\alias{RiscoRelativo}
\title{Função Risco Relativo}
\usage{
RiscoRelativo(tabela, alpha = 0.05, referencerow = 2)
}
\arguments{
\item{tabela}{é uma tabela de tabulação cruzada. A tabela 2 x 2 deve ser criada
antes e nomeada como tabela}

\item{alpha}{é o nível de significância, padrão é 0.05}

\item{referencerow}{é a linha da tabela que será a referência, o padrão é 2}
}
\description{
Esta função calcula o Risco Relativo a partir de uma tabela 2 x 2
}
\details{
Basta criar uma tabela 2 x 2 para calcular RR, usando o formato:

tabela <- matrix(c(valor1, valor2, valor3, valor4), nrow=2, byrow=TRUE)

colnames(tabela) <- c("Doença","Sem doença")

rownames(tabela) <- c("Exposto","Não exposto")

print(tabela)

Usar a tabela no modelo
}
\author{
Petronio Fagundes de Oliveira Filho
}
