% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarioGrupos.R
\name{sumarioGrupos}
\alias{sumarioGrupos}
\title{Função que calcula resumo de vários grupos}
\usage{
sumarioGrupos(dados = NULL, var.grupo = NULL, var.teste = NULL)
}
\arguments{
\item{var.grupo}{é uma variável categórica que define os grupos}

\item{data}{é um dataframe com os dados}

\item{y}{é a variável numérica que será resumida}
}
\description{
Esta função calcula um resumo de vários grupos
}
